Move matrix_solver_3.py to scripts
